+++ organization/src/epilot/sdkconfiguration.py
@@ -19,9 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.46.1'
-    gen_version: str = '2.143.2'
-    user_agent: str = 'speakeasy-sdk/python 1.46.1 2.143.2 1.0.0 epilot-organization'
+    sdk_version: str = '1.46.2'
+    gen_version: str = '2.147.0'
+    user_agent: str = 'speakeasy-sdk/python 1.46.2 2.147.0 1.0.0 epilot-organization'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: